Fixes to support building in WebGL mode. - Moved editors into Editor directory - Forced loading of Singleton scriptableobjects in Title scene - Forced 1920x1080 resolution
