+++ Assets/Editor/Array2DEditorEnemySpawnDefinition.cs
@@ -1,0 +1,5 @@
+﻿using UnityEngine;
+using UnityEditor;
+
+[CustomEditor(typeof(Array2DEnemySpawnDefinition))]
+public class Array2DEditorEnemySpawnDefinition : Array2DEditor<EnemyTypeDefinition> { }
