+++ Assets/Rooms/Spawn/Array2DEnemySpawnDefinition.cs
@@ -1,10 +1,5 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEditor;
-
 
 public class Array2DEnemySpawnDefinition : Array2D<EnemyTypeDefinition> { }
-
-[CustomEditor(typeof(Array2DEnemySpawnDefinition))]
-public class Array2DEditorEnemySpawnDefinition : Array2DEditor<EnemyTypeDefinition> { }
